PkCreative/requirements.txt
beautifulsoup4==4.6.0
captcha==0.3
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.9
django-allauth==0.41.0
django-cogwheels==0.3
django-modelcluster==4.4
django-recaptcha==2.0.5
django-taggit==0.24.0
django-treebeard==4.3.1
django-widget-tweaks==1.4.5
djangorestframework==3.11.0
draftjs-exporter==2.1.7
gunicorn==20.0.4
html5lib==1.0.1
idna==2.8
oauthlib==3.1.0
Pillow==6.2.1
psycopg2-binary==2.8.4
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
six==1.13.0
sqlparse==0.3.0
Unidecode==1.1.1
urllib3==1.25.7
wagtail==2.6.3
wagtail-django-recaptcha==1.0
wagtailfontawesome==1.1.4
wagtailmenus==3.0
wagtailnews==2.7.0
wagtailstreamforms==3.8
webencodings==0.5.1
Willow==1.1
